+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 03-02-2021
+### Added
+- Permissions for the DSMR API
+- Stored procedure to compute energy usage/production
+
+### Updated
+- Documentation
+- README file
+- Security/support definition
+
 ## [1.5.0] - 27-04-2021
 ### Added
 - Stored procedure to group power data by hour